requirements.txt
botocore==1.35.10
exceptiongroup==1.2.2
iniconfig==2.0.0
jmespath==1.0.1
mysql-connector-python==9.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
python-dateutil==2.9.0.post0
six==1.16.0
tomli==2.0.1
urllib3==1.26.20
aws_secretsmanager_caching==1.1.2
python-dotenv==1.0.1
